projects/GoogleCalendar/requirements.txt
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-WTF==0.14.3
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.1
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.3.1
